Only react to pvc command if run by the room owner

+++ model/private-vc.js
@@ -71,7 +71,7 @@ const PrivateVC = {
             );
 
             if (requestor && requestor !== member.id) {
-                PrivateVC.privateVoiceChatJoinHandler(requestor, newVoiceState);
+                PrivateVC.privateVoiceChatJoinHandler(requestor, newVoiceState.member);
             }
         }
 
@@ -372,6 +372,13 @@ const PrivateVC = {
      * @param {VoiceState} oldVoiceState
      */
     privateVoiceChatRequestHandler: async (member, oldVoiceState) => {
+        if (PrivateVC.list[member.id]) {
+            await Guild.botChannel.send(trans('model.privateVC.errors.alreadyExists.mods', [member.toString()], 'en'));
+            await member.send(trans('model.privateVC.errors.alreadyExists.member'));
+            await member.voice.setChannel(null);
+            return;
+        }
+
         const voiceChannelsCount = Guild.smallVoiceCategoryChannel.children.size;
         let textChannel, voiceChannel, waitingChannel;
 
@@ -475,14 +482,11 @@ const PrivateVC = {
 
     /**
      * @param {Snowflake} requestor
-     * @param {VoiceState} newVoiceState
+     * @param {GuildMember} guestMember
      */
-    privateVoiceChatJoinHandler: async (requestor, newVoiceState) => {
+    privateVoiceChatJoinHandler: async (requestor, guestMember) => {
         const hostMember = Guild.discordGuild.member(requestor);
-
-        const guestMember = newVoiceState.member;
         const guestUser = guestMember.user;
-
         const emojis = ['pollyes', 'pollno'].map(name => bot.emojis.cache.find(emoji => emoji.name === name));
         const channels = PrivateVC.list[requestor].slice(0, 3).map(id => Guild.discordGuild.channels.cache.find(channel => channel.id === id));
 
@@ -536,6 +540,7 @@ const PrivateVC = {
 
     /**
      * @param {GuildMember} currentHostMember
+     * @returns {GuildMember}
      */
     privateVoiceChatPropertyTransferHandler: async (currentHostMember, channels) => {
         const memberPermissionLevels = new Discord.Collection();
@@ -590,6 +595,8 @@ const PrivateVC = {
 
             channels.filter(channel => channel !== undefined).forEach(channel => channel.delete());
         }
+
+        return newHostMember;
     },
 
     /**
@@ -598,6 +605,10 @@ const PrivateVC = {
      * @returns {bool}
      */
     setChannelUserLimit: async (hostMember, args) => {
+        if (!PrivateVC.list[hostMember.id]) {
+            return false;
+        }
+
         const channels = PrivateVC.list[hostMember.id].slice(0, 3).map(
             id => Guild.discordGuild.channels.cache.find(channel => channel.id === id)
         );
@@ -676,7 +687,7 @@ const PrivateVC = {
      * @param {GuildMember} hostMember
      * @param {Array} channels
      */
-    synchronizeChannels: (hostMember, channels) => {
+    synchronizeChannels: async (hostMember, channels) => {
         if (channels[1]) {
             if (channels[1].members.size === 0) {
                 // No one left in the voice channel: delete on-demand VC
@@ -684,7 +695,12 @@ const PrivateVC = {
             } else {
                 if (!channels[1].members.has(hostMember.id)) {
                     // Host member left the voice channel: transfer property
-                    PrivateVC.privateVoiceChatPropertyTransferHandler(hostMember, channels);
+                    const newHostMember = await PrivateVC.privateVoiceChatPropertyTransferHandler(hostMember, channels);
+
+                    if (channels[2]) {
+                        // Late join request handling of members who joined waiting room during channel synchronization grace period
+                        channels[2].members.each(member => PrivateVC.privateVoiceChatJoinHandler(newHostMember.id, member));
+                    }
                 }
 
                 PrivateVC.fixChannelPermissions(channels);
@@ -729,6 +745,10 @@ const PrivateVC = {
      * @param {string} name
      */
     renameChannels: async (hostMember, name) => {
+        if (!PrivateVC.list[hostMember.id]) {
+            return false;
+        }
+
         const channels = PrivateVC.list[hostMember.id].slice(0, 3).map(
             id => Guild.discordGuild.channels.cache.find(channel => channel.id === id)
         );
@@ -739,13 +759,15 @@ const PrivateVC = {
             Logger.exception(exception);
             await Guild.botChannel.send(trans('model.privateVC.errors.renameFailed.mods', [hostMember.toString()], 'en'));
             await channels[0].send(trans('model.privateVC.errors.renameFailed.member'));
-            return;
+            return false;
         }
 
         await Promise.all([
             channels[0].setName(name),
             channels[1].setName(name),
         ]);
+
+        return true;
     },
 
     /**
@@ -770,7 +792,7 @@ const PrivateVC = {
     waitingRoomLeaveHandler: async (guestMember, requestor) => {
         const channels = PrivateVC.list[requestor].slice(0, 3).map(id => Guild.discordGuild.channels.cache.find(channel => channel.id === id));
 
-        if (channels[1]) {
+        if (channels[1] && PrivateVC.pendingJoinRequests[channels[1].id]) {
             const message = PrivateVC.pendingJoinRequests[channels[1].id][guestMember.id];
 
             await message.delete();
